PyETL/requirements.txt
pandas
sqlalchemy
pyyaml
cryptography
psycopg2-binary
